requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
lifelines==0.27.8
xgboost==2.0.3
shap==0.45.0
mapie==0.8.3
fairlearn==0.10.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
faker==24.11.0
scipy==1.13.0
joblib==1.4.2
python-dateutil==2.9.0
google-genai==0.3.0


